backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
Pillow
python-decouple
psycopg2-binary
celery
redis
django-filter
django-extensions
whitenoise
gunicorn
